requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.6
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
numpy>=1.23
regex>=2023.12.25
spacy>=3.7.4
pydantic>=2.7
python-dotenv>=1.0.1
openai>=1.42.0
spacy>=3.7.4
